PIDMR-231 Split App component into smaller components, add Header separately
